requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-Moment==0.9.0
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==4.6.7
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
monotonic==1.5
psycopg2==2.8.4
python-dotenv==0.10.5
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.12
urllib3==1.25.7
vine==1.3.0
Werkzeug==0.16.0
WTForms==2.2.1
zipp==2.1.0
